Switched to new data file w/ updated wt-at-age; doubled SSB to compare
Found out that CEATTLE is essentially underestimating SSB by 1/2, which leads me to believe that the SSB calculations (which are based on numbers and weight-at-age) are not correct, rather than the data inputs from the assessment model being wrong. Need to fix that, along with the mismatch between age 0 and age 1.