Trying to get diet data ready for Dirichlet, but getting confused
There are multiple records for prey hake for some predator IDs, but the prey weights are the same. Is that because the prey weight is summed across that species for that stomach? Or is that a mistake from somewhere else in the code? Also, I don't trust the calculations. I seem to be missing some data.